requirements.txt
streamlit>=1.29.0
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.36.0
torch>=2.6.0
opencv-python>=4.8.0
deepface>=0.0.79
google-generativeai>=0.3.2
pandas>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
